+++ R/multinom_reg_data.R
@@ -234,12 +234,10 @@ set_pred(
   type = "class",
   value = list(
     pre = NULL,
-    post = function(x, object) {
-      object$lvl[x + 1]
-    },
-    func = c(pkg = "keras", fun = "predict_classes"),
+    post = NULL,
+    func = c(pkg = "parsnip", fun = "keras_predict_classes"),
     args =
-      list(object = quote(object$fit),
+      list(object = quote(object),
            x = quote(as.matrix(new_data)))
   )
 )
@@ -256,7 +254,7 @@ set_pred(
       x <- as_tibble(x)
       x
     },
-    func = c(pkg = "keras", fun = "predict_proba"),
+    func = c(fun = "predict"),
     args =
       list(object = quote(object$fit),
            x = quote(as.matrix(new_data)))