+++ tests/testthat/test_multinom_reg_keras.R
@@ -36,8 +36,10 @@ ctrl <- control_parsnip(verbosity = 0, catch = FALSE)
 test_that('model fitting', {
   skip_on_cran()
   skip_if_not_installed("keras")
+  skip_if(is.null(tensorflow::tf_version()))
 
-  set.seed(257)
+  set_tf_seed(257)
+
   expect_error(
     fit1 <-
       fit_xy(
@@ -49,7 +51,8 @@ test_that('model fitting', {
     regexp = NA
   )
 
-  set.seed(257)
+  set_tf_seed(257)
+
   expect_error(
     fit2 <-
       fit_xy(
@@ -103,6 +106,7 @@ test_that('model fitting', {
 test_that('classification prediction', {
   skip_on_cran()
   skip_if_not_installed("keras")
+  skip_if(is.null(tensorflow::tf_version()))
 
   library(keras)
 
@@ -147,10 +151,12 @@ test_that('classification prediction', {
 test_that('classification probabilities', {
   skip_on_cran()
   skip_if_not_installed("keras")
+  skip_if(is.null(tensorflow::tf_version()))
 
   library(keras)
 
-  set.seed(257)
+  set_tf_seed(257)
+
   lr_fit <-
     fit_xy(
       basic_mod,
@@ -160,14 +166,15 @@ test_that('classification probabilities', {
     )
 
   keras_pred <-
-    keras::predict_proba(lr_fit$fit, as.matrix(te_dat[, -5])) %>%
+    predict(lr_fit$fit, as.matrix(te_dat[, -5])) %>%
     as_tibble(.name_repair = "minimal") %>%
     setNames(paste0(".pred_", lr_fit$lvl))
 
   parsnip_pred <- predict(lr_fit, te_dat[, -5], type = "prob")
   expect_equal(as.data.frame(keras_pred), as.data.frame(parsnip_pred))
 
-  set.seed(257)
+  set_tf_seed(257)
+
   plrfit <-
     fit_xy(
       reg_mod,
@@ -177,7 +184,7 @@ test_that('classification probabilities', {
     )
 
   keras_pred <-
-    keras::predict_proba(plrfit$fit, as.matrix(te_dat[, -5])) %>%
+    predict(plrfit$fit, as.matrix(te_dat[, -5])) %>%
     as_tibble(.name_repair = "minimal") %>%
     setNames(paste0(".pred_", lr_fit$lvl))
   parsnip_pred <- predict(plrfit, te_dat[, -5], type = "prob")